Recycle Views/Status Bar/Capture Events
Spent the beginning of the day working with the status bar, but I ended
up moving on to implementing the two recycler views for the chats list
and the friends list. I have the friends list populated with fake data
and the chats list still needs and adapter. I also added different
event handlers for taking a picture and eventually video. The functions
don’t do anything right now, but they are called at the correct times.