Friend System Update/Visual Update
Finished the client side code for the friend system and updated the
visuals with new icons. I also fixed some bug that had to do with the
camera. I wasn’t able to get double tapping to work in a timely manner,
so I added a camera switch button. Finally, I added contrasting
backgrounds for the buttons in the camera.